Removing the unneeded checks in wannier_prog for plotting
These were used to decide hether to enter the plotting routines.
This was causing the `write_u_matrices` to be ignored if there
wasn't also another plotting flag active.
Now the routine is always called, and inside it there were already
the appropriate if statements.
This new design should be more future-proof when more plotting
modes are added. I only put the header printing within an if,
this is in the same file so easier to notice in the future, and
anyway even if forgotten will just not print the header.

This commit fixes #139