In berry module, modified name of output files for kdotp option to match existing format. Modified example in testsuite accordingly
Included options and information on kdotp option into the user guide.

Also, included description of shift current routine in "Overview of the Berry module" Chapter of the user guide, this was not done
before and went unnoticed. Also, moved description of input variables of shift current to "berry" section, these were misplaced in
"Global variables" section.