differentiate between mandatory lib args and opts
this involves splitting the input parser in two: one part reads
(possibly repeatedly) the .win file resetting simple options.
This allows, in the library context, the .win file to be used to
override default values for control parameters.

The second set of variables are settings which must logically be
provided, in the library context, by the calling code.  These include
the atomic positions, the number of wannier functions, etc, these cannot
be reset because the determine allocations in the library data struct.